+++ screens/HelpScreen.tsx
@@ -29,6 +29,7 @@ const HelpScreen = () => (
     </ScrollView>
 );
 
+// Place holder styles until decision is made on this pages implementation
 const styles = StyleSheet.create({
     container: {
         padding: 20,
